+++ src/main/java/com/DevPointSystem/Comptabilite/Recette/domaine/TransfertCaisse.java
@@ -102,7 +102,7 @@ public class TransfertCaisse {
     private Integer codeEtatApprouver;
 
     @Column(name = "code_user_approuver", columnDefinition = "Nvarchar(200) default ''")
-    private Integer codeUserApprouver;
+    private Long codeUserApprouver;
 
     @Temporal(TemporalType.TIMESTAMP)
     @Column(name = "date_approuve", columnDefinition = "datetime default(getdate())")
@@ -240,11 +240,11 @@ public class TransfertCaisse {
         this.codeEtatApprouver = codeEtatApprouver;
     }
 
-    public Integer getCodeUserApprouver() {
+    public Long getCodeUserApprouver() {
         return codeUserApprouver;
     }
 
-    public void setCodeUserApprouver(Integer codeUserApprouver) {
+    public void setCodeUserApprouver(Long codeUserApprouver) {
         this.codeUserApprouver = codeUserApprouver;
     }
