+++ src/main/java/com/DevPointSystem/Comptabilite/Recette/dto/TransfertCaisseDTO.java
@@ -60,7 +60,7 @@ public class TransfertCaisseDTO {
 
     private Integer codeEtatApprouver;
 
-    private Integer codeUserApprouver;
+    private Long codeUserApprouver;
 
     private Date dateApprouve;
 
@@ -193,11 +193,11 @@ public class TransfertCaisseDTO {
         this.codeEtatApprouver = codeEtatApprouver;
     }
 
-    public Integer getCodeUserApprouver() {
+    public Long getCodeUserApprouver() {
         return codeUserApprouver;
     }
 
-    public void setCodeUserApprouver(Integer codeUserApprouver) {
+    public void setCodeUserApprouver(Long codeUserApprouver) {
         this.codeUserApprouver = codeUserApprouver;
     }
